backend/requirements.txt
fastapi
uvicorn
openai
deepgram-sdk
httpx
python-dotenv
